Make package buildable as a module: "zappem.net/pub/net/fastping"
This is a clone of https://github.com/tatsushid/go-fastping.git as of
the date of this commit. I wanted to use the code for a project and
needed it to build as a module using a recent toolchain. This remains
licensed under the same MIT terms as the original sources (see LICENSE).

To get the sources to test OK I've fixed a formatting error in one test
log line, and replaced the use of 127.0.0.100 as an unpingable IP4
address, with 192.0.2.1. The former, is expected to be pingable
according to:

  https://networkengineering.stackexchange.com/a/44636

However, https://en.wikipedia.org/wiki/Reserved_IP_addresses suggests
the 192.0.2.1 replacement address is reserved for some documentation
purposes, so I'm thinking it is unlikely to be pingable on most systems.

Signed-off-by: Tinkerer <tinkerer@zappem.net>